Update the landing page of the Recipes index page
Use CSS to update the home page of the recipes website and make it
look a bit more appealing, with the look of a card instead of a list.
Change the background colors to purple and the recipe elements to be
anchors instead of list items, and they are no longer wrapped in an
unordered list.